+++ Assets/AXIP/AILIA-MODELS/AudioProcessing/AiliaAudioProcessingSample.cs
@@ -23,6 +23,8 @@ namespace ailiaSDK {
 			rvc,
 			rvc_with_f0,
 			whisper_tiny,
+			whisper_small,
+			whisper_medium,
 		}
 
 		[SerializeField]
@@ -57,15 +59,13 @@ namespace ailiaSDK {
 		private long rvc_time = 0;
 		private long f0_time = 0;
 
-		//Preview
-		private Texture2D wave_texture = null;
-
 		//AILIA
 		private AiliaSileroVad ailia_vad = new AiliaSileroVad();
 		private AiliaRvc ailia_rvc = new AiliaRvc();
 		private AiliaMicrophone ailia_mic = new AiliaMicrophone();
 		private AiliaSplitAudio ailia_split = new AiliaSplitAudio();
 		private AiliaSpeechModel ailia_speech = new AiliaSpeechModel();
+		private AiliaDisplayAudio ailia_display_audio = new AiliaDisplayAudio();
 
 		//AILIA open file
 		private AiliaDownload ailia_download = new AiliaDownload();
@@ -149,27 +149,39 @@ namespace ailiaSDK {
 					}));
 					break;
 				case AudioProcessingModels.whisper_tiny:
+				case AudioProcessingModels.whisper_small:
+				case AudioProcessingModels.whisper_medium:
 					mode_text.text = "whisper";
 
-					string encoder_path = "encoder_tiny.opt3.onnx";
-					string decoder_path = "decoder_tiny_fix_kv_cache.opt3.onnx";
+					string encoder_path = "";
+					string decoder_path = "";
 					string vad_path = "silero_vad.onnx";
 
-					urlList.Add(new ModelDownloadURL() { folder_path = "silero-vad", file_name = "silero_vad.onnx" });
-					urlList.Add(new ModelDownloadURL() { folder_path = "whisper", file_name = encoder_path });
-					urlList.Add(new ModelDownloadURL() { folder_path = "whisper", file_name = decoder_path });
-
 					int task = AiliaSpeech.AILIA_SPEECH_TASK_TRANSCRIBE; //AiliaSpeech.AILIA_SPEECH_TASK_TRANSLATE;
-					int flag = AiliaSpeech.AILIA_SPEECH_FLAG_NONE; //AiliaSpeech.AILIA_SPEECH_FLAG_LIVE;
+					int flag = AiliaSpeech.AILIA_SPEECH_FLAG_LIVE; //AiliaSpeech.AILIA_SPEECH_FLAG_NONE;
 					int memory_mode = Ailia.AILIA_MEMORY_REDUCE_CONSTANT | Ailia.AILIA_MEMORY_REDUCE_CONSTANT_WITH_INPUT_INITIALIZER | Ailia.AILIA_MEMORY_REUSE_INTERSTAGE;
 					int env_id = GetEnvId(gpu_mode);
-					int api_model_type = AiliaSpeech.AILIA_SPEECH_MODEL_TYPE_WHISPER_MULTILINGUAL_TINY;
+					int api_model_type = 0;
+					if (ailiaModelType == AudioProcessingModels.whisper_tiny){
+						api_model_type = AiliaSpeech.AILIA_SPEECH_MODEL_TYPE_WHISPER_MULTILINGUAL_TINY;
+						encoder_path = "encoder_tiny.opt3.onnx";
+						decoder_path = "decoder_tiny_fix_kv_cache.opt3.onnx";
+					}
+					if (ailiaModelType == AudioProcessingModels.whisper_small){
+						api_model_type = AiliaSpeech.AILIA_SPEECH_MODEL_TYPE_WHISPER_MULTILINGUAL_SMALL;
+						encoder_path = "encoder_small.opt3.onnx";
+						decoder_path = "decoder_small_fix_kv_cache.opt3.onnx";
+					}
+					if (ailiaModelType == AudioProcessingModels.whisper_medium){
+						api_model_type = AiliaSpeech.AILIA_SPEECH_MODEL_TYPE_WHISPER_MULTILINGUAL_MEDIUM;
+						encoder_path = "encoder_medium.opt3.onnx";
+						decoder_path = "decoder_medium_fix_kv_cache.opt3.onnx";
+					}
+					urlList.Add(new ModelDownloadURL() { folder_path = "silero-vad", file_name = vad_path });
+					urlList.Add(new ModelDownloadURL() { folder_path = "whisper", file_name = encoder_path });
+					urlList.Add(new ModelDownloadURL() { folder_path = "whisper", file_name = decoder_path });
 					bool virtual_memory_enable = false;
 					string language = "auto"; // ja
-					if (virtual_memory_enable){
-						Ailia.ailiaSetTemporaryCachePath(Application.temporaryCachePath);
-						memory_mode = Ailia.AILIA_MEMORY_REDUCE_CONSTANT | Ailia.AILIA_MEMORY_REDUCE_CONSTANT_WITH_INPUT_INITIALIZER | Ailia.AILIA_MEMORY_REUSE_INTERSTAGE | Ailia.AILIA_MEMORY_REDUCE_CONSTANT_WITH_FILE_MAPPED;
-					}
 					StartCoroutine(ailia_download.DownloadWithProgressFromURL(urlList, () =>
 					{
 						FileOpened = ailia_speech.Open(asset_path + "/" + encoder_path, asset_path + "/" + decoder_path, env_id, memory_mode, api_model_type, task, flag, language);
@@ -211,84 +223,8 @@ namespace ailiaSDK {
 		{
 			ailia_vad.Close();
 			ailia_rvc.Close();
+			ailia_speech.Close();
 		}		
-
-		// Display pcm and vad confidence value
-		private float[] displayWaveData = null;
-		private float[] displayConfData = null;
-
-		private void DisplayPreviewPcm(Color32 [] colors, float [] waveData, float [] conf, uint channels){
-			int steps = 10;
-			int w = wave_texture.width;
-			int h = wave_texture.height / 4;
-			int original_h = wave_texture.height;
-			int buf_w = w * steps;
-			int offset_y = 3 * h;
-
-			if (displayWaveData == null){
-				displayWaveData = new float[buf_w];
-				displayConfData = new float[buf_w];
-			}
-
-			int add_data_n = (int)(waveData.Length / channels);
-			int reuse_data_n = buf_w - add_data_n;
-			for (int i = 0; i < reuse_data_n; i++){
-				displayWaveData[i] = displayWaveData[i + (buf_w - reuse_data_n)];
-				displayConfData[i] = displayConfData[i + (buf_w - reuse_data_n)];
-			}
-			for (int i = reuse_data_n; i < buf_w; i++){
-				if (i >= 0){
-					displayWaveData[i] = waveData[(i - reuse_data_n) * channels];
-					if (conf == null){
-						displayConfData[i] = 0.0f;
-					}else{
-						displayConfData[i] = conf[(i - reuse_data_n) * channels];
-					}
-				}
-			}
-
-			for (int x = 0; x < w ; x++){
-				for (int y = 0; y < original_h ; y++){
-					colors[y*w+x] = new Color32(0,0,0,255);
-				}
-				int y3 = (int)(displayConfData[x * steps] * h);
-				if (y3 >= 0 && y3 < h){
-					for (int y = 0; y < y3 ; y++){
-						colors[(y+offset_y)*w+x] = new Color32(255,0,0,255);
-					}
-				}
-				int y2 = (int)(displayWaveData[x * steps] * h / 2 + h / 2);
-				if (y2 >= 0 && y2 < h){
-					colors[(y2+offset_y)*w+x] = new Color32(0,255,0,255);
-				}
-			}
-		}
-
-		// Display splitted audio clip
-		private void DisplayVadAudioClip(Color32 [] colors, AudioClip clip, int i, int count){
-			float [] buf = new float[clip.channels * clip.samples];
-			clip.GetData(buf, 0);
-
-			int w = wave_texture.width;
-			int div = count;
-			if (div < 3){
-				div = 3;
-			}
-			int h = (wave_texture.height - wave_texture.height/4) / div;
-			int steps = clip.samples / w;
-			if (steps < 1){
-				steps = 1;
-			}
-			int offset_y = 3 * wave_texture.height/4 - (i + 1) * h;
-			int buf_w = w * steps;
-
-			for (int x = 0; x < w ; x++){
-				int y2 = (int)(buf[x * steps] * h / 2 + h / 2);
-				if (y2 >= 0 && y2 < h){
-					colors[(offset_y + y2)*w+x] = new Color32(0,255,0,255);
-				}
-			}
-		}
 
 		// Use this for initialization
 		void Start()
@@ -310,13 +246,8 @@ namespace ailiaSDK {
 			Clear();
 
 			//Get camera image
-			int tex_width = 480;
-			int tex_height = 480;
-			if (wave_texture == null)
-			{
-				wave_texture = new Texture2D(tex_width, tex_height);
-				raw_image.texture = wave_texture;
-			}
+			raw_image.texture = ailia_display_audio.Create();
+
 			
 			// Get Mic Input
 			float[] waveData = null;
@@ -324,43 +255,36 @@ namespace ailiaSDK {
 			uint frequency = 1;
 			waveData = ailia_mic.GetPcm(ref channels, ref frequency);
 
-			if (ailiaModelType == AudioProcessingModels.whisper_tiny){
+			if (ailiaModelType == AudioProcessingModels.whisper_tiny || ailiaModelType == AudioProcessingModels.whisper_small || ailiaModelType == AudioProcessingModels.whisper_medium){
 				WhisperUpdate(waveData, channels, frequency);
 			} else {
 				VadAndRvcUpdate(waveData, channels, frequency);
 			}
 		}
 
-		private List<float[]> waveQueue = new List<float[]>();
-
 		void WhisperUpdate(float[] waveData, uint channels, uint frequency){
-			// Add to queue
-			waveQueue.Add(waveData);
-
 			// Preview
-			Color32 [] colors = wave_texture.GetPixels32();
-			DisplayPreviewPcm(colors, waveData, null, channels);
-			wave_texture.SetPixels32(colors);
-			wave_texture.Apply();
+			ailia_display_audio.DisplayPcm(waveData, null, channels);
 
 			// Error handle
 			if (ailia_speech.IsError()){
 				return;
 			}
 
+			// Display Processing
+			if (ailia_speech.IsProcessing()){
+				if (content_text == ""){
+					content_text = "[processing]";
+				}
+			}
+
 			// Get result
 			WhisperDisplayIntermediateResult();
 			WhisperGetResult();
 
-			// Check processing
-			if (ailia_speech.IsProcessing()){
-				return;
-			}
-
-			// Transcribe from queue
-			bool complete = false;
-			ailia_speech.Transcribe(waveQueue, frequency, channels, complete);
-			waveQueue = new List<float[]>();
+			// Transcribe request
+			bool complete = ailia_mic.IsComplete();
+			ailia_speech.Transcribe(waveData, frequency, channels, complete);
 		}
 
 		string content_text = "";
@@ -369,7 +293,7 @@ namespace ailiaSDK {
 			string intermediateText = ailia_speech.GetIntermediateText();
 			if (content_text != "" || intermediateText != ""){
 				if (intermediateText != ""){
-					label_text.text = content_text + "[processing] " + intermediateText;
+					label_text.text = "[processing] " + intermediateText;
 				}else{
 					label_text.text = content_text;
 				}
@@ -378,11 +302,14 @@ namespace ailiaSDK {
 
 
 		private void WhisperGetResult(){
+			if (content_text == "[processing]"){
+				content_text = "";
+			}
 			List<string> results = ailia_speech.GetResults();
 			for (uint idx = 0; idx < results.Count; idx++){
 				string text = results[(int)idx];
 				string display_text = text + "\n";
-				content_text = content_text + display_text;
+				content_text = display_text + content_text;
 			}
 		}
 
@@ -432,13 +359,7 @@ namespace ailiaSDK {
 			}
 			
 			// Preview
-			Color32 [] colors = wave_texture.GetPixels32();
-			DisplayPreviewPcm(colors, vad_result.pcm, vad_result.conf, channels);
-			for (int i = 0; i < vad_audio_clip.Count; i++){
-				DisplayVadAudioClip(colors, vad_audio_clip[i], i, vad_audio_clip.Count);
-			}
-			wave_texture.SetPixels32(colors);
-			wave_texture.Apply();
+			ailia_display_audio.DisplayVad(vad_result, vad_audio_clip, channels);
 		}
 
 		private void RvcPushSplitAudio(AudioClip clip)
