+++ Assets/AXIP/AILIA-MODELS/AudioProcessing/AiliaDisplayAudio.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54c506d56ed4c4b36ab8d3b504cc4540
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
